fix: make all static pages offline
