+++ middleware.public.js
@@ -9,9 +9,6 @@ export const config = {
 export async function middleware(request) {
   const url = request.nextUrl;
 
-  url.pathname = '/api/v1/_responses/rate-limit-reached';
-  return NextResponse.rewrite(url);
-
   try {
     const rateLimitResult = await rateLimit.check(request);
