+++ pages/api_offline/v1/activation/index.public.js
