+++ pages/api_offline/v1/contents/[username]/[slug]/index.public.js
